requirements.txt
aeskeyschedule==0.0.3
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.3.1
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
Babel==2.8.0
beautifulsoup4==4.9.1
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
coverage==5.1
coveralls==2.0.0
cryptography==2.9.2
Cython==0.29.21
daphne==3.0.1
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.27.0
django-appconf==1.0.4
django-clear-cache==0.3
django-clearcache==1.0.1
django-cors-headers==3.3.0
django-cryptography==1.0
django-filter==2.4.0
django-gsheets==0.0.10
django-mirage-field==1.0.1
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-redis==4.12.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-tagging==0.5.0
django-taggit==1.3.0
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
django-user-visit==0.4.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
docopt==0.6.2
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
hiredis==1.0.1
html2text==2020.1.16
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
msgpack==1.0.0
msgpack-python==0.5.6
# mysqlclient @ file:///C:/Users/ike/Documents/coding/Python%20Project/backend_finger/mysqlclient-1.4.6-cp38-cp38-win32.whl
nose==1.3.7
numpy==1.19.0
oauthlib==3.1.0
opencv-contrib-python==4.3.0.36
packaging==20.5
phonenumbers==8.12.5
Pillow==7.1.2
protobuf==3.13.0
psycopg2==2.8.5
py3rijndael==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
# pygame==1.9.6
PyHamcrest==2.0.2
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3.1
redis==2.10.6
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.1
service-identity==18.1.0
setuptools-scm==4.1.2
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
tqdm==4.46.0
# Twisted @ file:///C:/Users/ike/Documents/coding/Python%20Project/github/django-channels-chat-master/Twisted-20.3.0-cp38-cp38-win32.whl
txaio==20.4.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.9
user-agents==2.2.0
webencodings==0.5.1
wrapt==1.11.2
zope.interface==5.1.0
